core: move patches to namespace
